implement gooCat Form
